Twins-detection/ai/requirements.txt
facenet-pytorch
Pillow
numpy